Fix for #205: "Users" menu visibility
This required adding a "superuser check" mixin to many of the view
classes.  That turned up a problem in django-tutelary that meant that
the superuser checks added in the mixin slowed the platform down a lot.
The tutelary bug was fixed and the tutelary version bumped here to take
account of that.