Set up to use wheelhouse for PIP packages
Use --find-links option to preferentially pull PIP packages from the
wheelhouse in the cadasta-wheelhouse S3 bucket (maintained by the code
in the Cadasta/wheelhouse repo).

For the moment, keeping the requirements files consistent between the
main repo and the wheelhouse repo has to be done manually, but this is
not a problem because PIP is conservative, in the sense that if it
doesn't find a package in the wheelhouse it will just download it via
the main package index.