Fix for show username in member permissions page (#687)
The issue https://github.com/Cadasta/cadasta-platform/issues/626 suggest that if full name is empty
then username should be set in member permission page. This pull request contains fix for it as well as
member list page, when full name is empty member column gets blank. So I include to display username if full name is not
present