Refactor functional tests user creation code
Reuse user creation code whether user is a superuser or not.  Apply the
superuser role if user is indeed supposed to be a superuser.

Display user's full name on user list page

Under the "member" column, it used to show only the user's username, now
it's showing "username (full name)".

Close #316