Implemented router to handle single page view calls. Added coordinates to url to allow user to store specific map locations. --- Router_mixin includes all functions that update the state of the current page.
Forms submitted in project detail are intercepted and handled through ajax to prevent redirects.

Custom headers raise form errors, which are then handled in router_mixins formSubmission

Custom headers raise permission errors, and redirects for anonymous and non-permissioned users and handled in router

Routes are provided with two functions: updatePage and updateState which determine what other events to fire.

The only times a map will move on automatically are when a location detail is reached through an outside url, or when first reaching a project page.

Added tab to url to redirect to specific tabs in location detail views.

Added catiline library to turn the blocking part of L.TileLayer.GeoJSON into a webworker